requirements.txt
<<<<<<< HEAD
fastapi
uvicorn[standard]
python-multipart
numpy==1.23.5
pillow
opencv-python-headless
tensorflow==2.12.0
keras==2.12.0
deepface==0.0.93
=======
fastapi
uvicorn[standard]
python-multipart
numpy==1.23.5
pillow
opencv-python-headless
tensorflow==2.12.0
keras==2.12.0
deepface==0.0.93
>>>>>>> 7597e87cea79753217e1414ef445050dfb860af7
